+++ core/components/Error.php
@@ -4,6 +4,7 @@ namespace Dev4Press\Plugin\CoreActivity\Components;
 
 use Dev4Press\Plugin\CoreActivity\Base\Component;
 use Dev4Press\Plugin\CoreActivity\Log\Core;
+use Dev4Press\v47\Core\Helpers\IP;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;